requirements.txt
fastapi
uvicorn[standard]
requests
google-genai
python-multipart
python-dotenv